requirements.txt
gradio==5.*
openai>=1.23.0
langchain-core>=0.2.0
langchain-community>=0.2.0
faiss-cpu~=1.7
python-dotenv>=1.0
